requirements.txt
aiogram==3.10.0
aiohttp==3.9.3
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.20.0
python-dotenv==1.0.0
pytz==2024.1
